add webview swift util for reload
