+++ apps/desktop/src-tauri/native/macos/Sources/sd-desktop-macos/webview.swift
@@ -1,0 +1,8 @@
+import WebKit
+
+@_cdecl("reload_webview")
+public func reloadWebview(webview: WKWebView) -> () {
+	webview.window!.orderOut(webview);
+	webview.reload();
+	webview.window!.makeKey();
+}
