+++ apps/desktop/src-tauri/src/macos/mod.rs
@@ -1,2 +1,7 @@
+mod native;
+
 mod window;
 pub use window::*;
+
+mod webview;
+pub use webview::*;
