+++ apps/desktop/src-tauri/src/macos/native.rs
@@ -1,0 +1,3 @@
+use std::ffi::c_void;
+
+pub type NSObject = *mut c_void;
