+++ apps/desktop/src-tauri/src/macos/webview.rs
@@ -1,0 +1,4 @@
+use super::native::NSObject;
+use swift_rs::*;
+
+pub_swift_fn!(reload_webview(webview: NSObject));
