+++ apps/desktop/src-tauri/src/macos/window.rs
@@ -1,7 +1,5 @@
-use std::ffi::c_void;
+use super::native::NSObject;
 use swift_rs::*;
-
-pub type NSObject = *mut c_void;
 
 pub_swift_fn!(lock_app_theme(theme_type: Int));
 pub_swift_fn!(blur_window_background(window: NSObject));
